Move interior generation into own file
